requirements.txt
Requests==2.31.0
slackclient==2.9.4
